Only add custom GRUB entries if not present

+++ arch-post-install.bash
@@ -291,17 +291,19 @@ install_bootloader()
 		sed -i "/GRUB_CMDLINE_LINUX_DEFAULT=/ c GRUB_CMDLINE_LINUX_DEFAULT=\"$kernel_params\"" /etc/default/grub
 
 		# Add custom GRUB entries
-		cat >> /etc/grub.d/40_custom <<-CUSTOM_GRUB
-			menuentry 'System shutdown' --class shutdown {
-			    echo 'System shutting down...'
-			    halt
-			}
+		if ! grep -i -q "menuentry" /etc/grub.d/40_custom; then
+			cat >> /etc/grub.d/40_custom <<-CUSTOM_GRUB
+				menuentry 'System shutdown' --class shutdown {
+				    echo 'System shutting down...'
+				    halt
+				}
 
-			menuentry 'System restart' --class restart {
-			    echo 'System rebooting...'
-			    reboot
-			}
-		CUSTOM_GRUB
+				menuentry 'System restart' --class restart {
+				    echo 'System rebooting...'
+				    reboot
+				}
+			CUSTOM_GRUB
+		fi
 
 		# Generate GRUB config file
 		print_progress_text "Generating GRUB config file"
